+++ frontend/src/mocks/MockIntersectionObserver.ts
