+++ frontend/src/setupTests.ts
@@ -1,4 +1,5 @@
-import MockIntersectionObserver from '__test__/MockIntersectionObserver';
+import MockIntersectionObserver from 'mocks/MockIntersectionObserver';
+import 'mocks/matchMedia';
 import { server } from 'mocks/server';
 
 import '@testing-library/jest-dom';
@@ -24,16 +25,16 @@ beforeEach(() => {
   globalThis.IS_REACT_ACT_ENVIRONMENT = true;
 });
 
-Object.defineProperty(window, 'matchMedia', {
-  writable: true,
-  value: jest.fn().mockImplementation((query) => ({
-    matches: false,
-    media: query,
-    onchange: null,
-    addListener: jest.fn(), // Deprecated
-    removeListener: jest.fn(), // Deprecated
-    addEventListener: jest.fn(),
-    removeEventListener: jest.fn(),
-    dispatchEvent: jest.fn(),
-  })),
-});
+// Object.defineProperty(window, 'matchMedia', {
+//   writable: true,
+//   value: jest.fn().mockImplementation((query) => ({
+//     matches: false,
+//     media: query,
+//     onchange: null,
+//     addListener: jest.fn(), // Deprecated
+//     removeListener: jest.fn(), // Deprecated
+//     addEventListener: jest.fn(),
+//     removeEventListener: jest.fn(),
+//     dispatchEvent: jest.fn(),
+//   })),
+// });
